requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
dj-database-url==0.5.0
Django==4.0.3
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-registration-redux==2.10
django-tinymce==3.4.0
djangorestframework==3.13.1
Pillow==9.0.1
psycopg2==2.9.3
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
whitenoise==6.0.0
